c-proj1/requirements.txt
anthropic==0.87.0
python-dotenv==1.2.2
pygame==2.6.1
